fix compatible with spring boot
