+++ runtime-sofa-boot-starter/src/main/resources/META-INF/spring.factories
@@ -4,6 +4,3 @@ com.alipay.sofa.runtime.spring.initializer.SofaRuntimeSpringContextInitializer
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.alipay.sofa.runtime.spring.configuration.SofaRuntimeAutoConfiguration
 
-
-org.springframework.context.ApplicationListener=\
-com.alipay.sofa.runtime.spring.listener.SofaRuntimeApplicationListener
