+++ runtime-sofa-boot-starter/src/test/java/com/alipay/sofa/runtime/ark/SofaEventHandlerTest.java
@@ -30,8 +30,7 @@ import com.alipay.sofa.runtime.spi.binding.Contract;
 import com.alipay.sofa.runtime.spi.component.SofaRuntimeContext;
 import com.alipay.sofa.runtime.spi.component.SofaRuntimeManager;
 import com.alipay.sofa.runtime.spi.service.ServiceProxy;
-import com.alipay.sofa.runtime.spring.configuration.SofaRuntimeAutoConfiguration;
-import com.alipay.sofa.runtime.spring.listener.SofaRuntimeApplicationListener;
+import com.alipay.sofa.runtime.spring.initializer.SofaRuntimeSpringContextInitializer;
 import mockit.Expectations;
 import mockit.Mock;
 import mockit.MockUp;
@@ -41,16 +40,15 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
-import org.springframework.boot.context.event.ApplicationPreparedEvent;
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
-
-import static org.mockito.Mockito.when;
 
 /**
  * @author qilong.zql
@@ -58,31 +56,25 @@ import static org.mockito.Mockito.when;
  */
 public class SofaEventHandlerTest {
 
-    private AnnotationConfigApplicationContext applicationContext;
+    private ConfigurableApplicationContext applicationContext;
 
     @Mocked
-    private Biz                                biz;
+    private Biz                            biz;
     @Mocked
-    private SofaRuntimeManager                 sofaRuntimeManager;
+    private SofaRuntimeManager             sofaRuntimeManager;
     @Mocked
-    private Contract                           contract;
+    private Contract                       contract;
     @Mocked
-    private MethodInvocation                   invocation;
+    private MethodInvocation               invocation;
 
     @Before
     public void before() {
-        applicationContext = new AnnotationConfigApplicationContext();
-        TestPropertyValues.of("com.alipay.sofa.boot.disableJvmFirst=true")
-            .and("com.alipay.sofa.boot.skipJvmReferenceHealthCheck=true")
-            .applyTo(applicationContext);
-        ApplicationPreparedEvent applicationPreparedEvent = Mockito
-            .mock(ApplicationPreparedEvent.class);
-        when(applicationPreparedEvent.getApplicationContext()).thenReturn(applicationContext);
-
-        this.applicationContext.register(SofaRuntimeAutoConfiguration.class,
-            SofaBootHealthCheckInitializer.class, XmlConfiguration.class);
-        new SofaRuntimeApplicationListener().onApplicationEvent(applicationPreparedEvent);
-        this.applicationContext.refresh();
+        Map<String, Object> properties = new HashMap<>();
+        properties.put("com.alipay.sofa.boot.disableJvmFirst", "true");
+        properties.put("com.alipay.sofa.boot.skipJvmReferenceHealthCheck", "true");
+        SpringApplication springApplication = new SpringApplication(XmlConfiguration.class);
+        springApplication.setDefaultProperties(properties);
+        this.applicationContext = springApplication.run(new String[] {});
     }
 
     @Test
@@ -179,9 +171,11 @@ public class SofaEventHandlerTest {
         for (SofaRuntimeManager runtimeManager : runtimeManagers) {
             SofaFramework.unRegisterSofaRuntimeManager(runtimeManager);
         }
+        SofaRuntimeSpringContextInitializer.setIsInitiated(false);
     }
 
     @Configuration
+    @EnableAutoConfiguration
     @ComponentScan("com.alipay.sofa.runtime.beans.impl")
     public static class XmlConfiguration {
     }
