+++ runtime-sofa-boot-starter/src/test/java/com/alipay/sofa/runtime/integration/TestSofaServiceAndReferenceException.java
@@ -24,6 +24,7 @@ import java.util.Map;
 import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
 import org.junit.Test;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 
 import com.alipay.sofa.runtime.api.annotation.SofaReference;
@@ -33,6 +34,7 @@ import com.alipay.sofa.runtime.beans.impl.SampleServiceImpl;
 import com.alipay.sofa.runtime.beans.service.SampleService;
 import com.alipay.sofa.runtime.integration.base.TestBase;
 import com.alipay.sofa.runtime.spring.bean.SofaBeanNameGenerator;
+import org.springframework.context.annotation.Configuration;
 
 /**
  * @author qilong.zql
@@ -70,6 +72,8 @@ public class TestSofaServiceAndReferenceException extends TestBase {
                                                SampleService.class, "")));
     }
 
+    @EnableAutoConfiguration
+    @Configuration
     static class TestSofaReferenceConfiguration {
         @Bean
         SampleService sampleService(@SofaReference(uniqueId = "rpc", binding = @SofaReferenceBinding(bindingType = "bolt")) SampleService sampleService) {
@@ -77,6 +81,8 @@ public class TestSofaServiceAndReferenceException extends TestBase {
         }
     }
 
+    @Configuration
+    @EnableAutoConfiguration
     static class TestSofaServiceConfiguration {
         @Bean
         @SofaService
